requirements.txt
Django>=5.0,<6.0
qdrant-client>=1.8,<2.0
python-dotenv>=1.0,<2.0
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
sentence-transformers>=2.6,<3.0
aiofiles>=23.0,<24.0
pillow>=10.0,<11.0